Implement functionality to play level
When starting a particular level, the user can see the items that should be found in the header. They can click on any area of the image and choose between items. If there is the chosen item in the area, the user will see a success alert and can continue looking for other items – the items list on click now contains only items that are not found. If not, the user will see a failure alert and can continue looking for items.